change day image
House씬 Canvas -> Day Image ctrl
Assemble2 오류 수정

+++ PBL_01/Assets/Scripts/Day_manager.cs
@@ -4,21 +4,23 @@ using UnityEngine;
 
 public class Day_manager : MonoBehaviour
 {
+    public GameObject day_image;
 
     int cnt = 0;
     // Start is called before the first frame update
     void Start()
     {
-        if(PlayerPrefs.GetInt("day", -1) == 1) //day1 일 때 조건 X
+        if(PlayerPrefs.GetInt("day", -1) == 1) //day1 일 때
         {
             PlayerPrefs.SetInt("sleep", 0); //잠자기
         }
         else
         {
             //초기화 되어서 다음날이 되었을 때
-            if (!PlayerPrefs.HasKey("bap") && !PlayerPrefs.HasKey("pill") && !PlayerPrefs.HasKey("planter")
-                && !PlayerPrefs.HasKey("random1") && !PlayerPrefs.HasKey("random2") && !PlayerPrefs.HasKey("sleep"))
+            if ((bool)GetBool("sleep"))
             {
+                day_image.GetComponent<day_Image_ctrl>().Day_Image();
+
                 PlayerPrefs.SetInt("bap", 0); //밥먹기
                 PlayerPrefs.SetInt("pill", 0); //약먹기
                 PlayerPrefs.SetInt("planter", 0); //화분 가꾸기
@@ -38,17 +40,20 @@ public class Day_manager : MonoBehaviour
         {
             if ((bool)GetBool("sleep"))
             {
-                PlayerPrefs.DeleteKey("sleep");
-
                 cnt += PlayerPrefs.GetInt("day") + 1;
                 PlayerPrefs.SetInt("day", cnt);
+
+                //PlayerPrefs.DeleteKey("sleep");
             }
         }
         else
         {
             //모두 진행되었을 때 키 삭제 초기화
-            if ((bool)GetBool("sleep"))
+            if ((bool)GetBool("sleep") && (bool)GetBool("bap") && (bool)GetBool("pill") && (bool)GetBool("planter")
+                        && (bool)GetBool("random1") && (bool)GetBool("random2") && (bool)GetBool("routine"))
             {
+                cnt += PlayerPrefs.GetInt("day") + 1;
+                PlayerPrefs.SetInt("day", cnt);
 
                 PlayerPrefs.DeleteKey("bap");
                 PlayerPrefs.DeleteKey("pill");
@@ -56,11 +61,8 @@ public class Day_manager : MonoBehaviour
                 PlayerPrefs.DeleteKey("random1");
                 PlayerPrefs.DeleteKey("random2");
                 PlayerPrefs.DeleteKey("routine");
-                PlayerPrefs.DeleteKey("sleep");
+                //PlayerPrefs.DeleteKey("sleep");
                 PlayerPrefs.DeleteKey("NoteCp");
-
-                cnt += PlayerPrefs.GetInt("day") + 1;
-                PlayerPrefs.SetInt("day", cnt);
             }
         }
     }
